requirements.txt
asgiref
autopep8
django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-dotenv
gunicorn
whitenoise
Pillow
python-decouple
django-middleware-global-request
jarowinkler #fuzzy search
djangorestframework
psycopg2-binary
django-crispy-forms
crispy-bootstrap5
requests
django-formtools